Begin prototyping active learning classifier. Doesn't run yet.

+++ zoobot/tfrecord/gz2_to_tfrecord.py
@@ -10,7 +10,7 @@ from tqdm import tqdm
 from zoobot.tfrecord import create_tfrecord
 
 
-def write_catalog_to_train_test_tfrecords(df, train_loc, test_loc, img_size, columns_to_save, train_test_fraction=0.8):
+def write_catalog_to_train_test_tfrecords(df, label_col, train_loc, test_loc, img_size, columns_to_save, train_test_fraction=0.8):
     train_test_split = int(train_test_fraction * len(df))
 
     df = df.sample(frac=1).reset_index(drop=True)
@@ -20,31 +20,32 @@ def write_catalog_to_train_test_tfrecords(df, train_loc, test_loc, img_size, col
     assert not train_df.empty
     assert not test_df.empty
 
-    write_image_df_to_tfrecord(train_df, train_loc, img_size, columns_to_save)
-    write_image_df_to_tfrecord(test_df, test_loc, img_size, columns_to_save)
+    write_image_df_to_tfrecord(train_df, label_col, train_loc, img_size, columns_to_save)
+    write_image_df_to_tfrecord(test_df, label_col, test_loc, img_size, columns_to_save)
+
+    return train_df, test_df
 
 
-def write_image_df_to_tfrecord(df, tfrecord_loc, img_size, columns_to_save):
+def write_image_df_to_tfrecord(df, label_col, tfrecord_loc, img_size, columns_to_save, append=False):
 
-    if os.path.exists(tfrecord_loc):
-        print('{} already exists - deleting'.format(tfrecord_loc))
-        os.remove(tfrecord_loc)
+    if not append:
+        if os.path.exists(tfrecord_loc):
+            print('{} already exists - deleting'.format(tfrecord_loc))
+            os.remove(tfrecord_loc)
 
     writer = tf.python_io.TFRecordWriter(tfrecord_loc)
 
     dimensions = img_size, img_size
 
     for subject_n, subject in tqdm(df.iterrows(), total=len(df), unit=' subjects saved'):
-        # print(subject_n)
         if subject['png_ready']:
             img = Image.open(subject['png_loc'])
-            img.thumbnail(dimensions)  # inplace on img
+            img.thumbnail(dimensions)  # inplace on img  # TODO split into fits/png. If fits, rescale directly
             matrix = np.array(img)
-            label = int(subject['t04_spiral_a08_spiral_weighted_fraction'])
+            label = int(subject[label_col])
             extra_data = {}
             for col in columns_to_save:
                 extra_data.update({col: subject[col]})
-            print(extra_data)
 
             create_tfrecord.image_to_tfrecord(matrix, label, writer, extra_data)
 
@@ -67,4 +68,5 @@ if __name__ == '__main__':
                          usecols=columns_to_save + ['png_loc', 'png_ready'],
                          nrows=None)
 
-        write_catalog_to_train_test_tfrecords(df, size, columns_to_save)
+        label_col = 't04_spiral_a08_spiral_weighted_fraction'
+        write_catalog_to_train_test_tfrecords(df, label_col, train_loc, test_loc, size, columns_to_save)
